requirements.txt
Django>=5.0
asgiref
sqlparse
tzdata
gunicorn
whitenoise[brotli]
dj-database-url
psycopg2-binary
PyJWT>=2.8.0
python-dotenv
